feat: removed NodeJS, purely static front-end code
