--- package.json
@@ -1,21 +1,0 @@
-{
-  "name": "sat-vis",
-  "version": "1.0.0",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "description": "",
-  "dependencies": {
-    "axios": "^1.7.9",
-    "cesium": "^1.125.0",
-    "cors": "^2.8.5",
-    "express": "^4.21.2",
-    "node-fetch": "^2.7.0",
-    "node-fetch-cache": "^5.0.2",
-    "satellite.js": "^5.0.0"
-  }
-}
